More-sophisticated cooling costs for SystemHeatBoiloff patch
The SystemHeatBoiloff patch was using the same CoolingHeatCost values
(per resource) for all parts, which meant that the CryoTanks ZBO tanks
had no advantage over regular tanks, despite being more expensive.  In
the basic non-SystemHeat configuration, the ZBO tanks have a lower
CoolingCost, so I've changed the SystemHeatPatch to compute the
CoolingHeatCost in proportion.  The ratio is such that non-ZBO tanks end
up with the same CoolingHeatCost values as before, and ZBO tanks get
better (since this won't break existing designs; doing it the other way
would mean some craft might no longer have enough cooling.)

The fusion fuels from Far Future Tech are a different story, since they
go in special tanks that don't have separate ZBO and non-ZBO versions,
and the ratio from CoolingCost to CoolingHeatCost is different than for
the chemical fuels.  The fixed CoolingHeatCost isn't a problem for
those, so I've left it as-is.